move utils.tracer to tests
