+++ tests/benchmark.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ddtrace import Tracer
-from tests.tracer.test_tracer import DummyWriter
+from tests import DummyWriter
 
 
 @pytest.fixture
