+++ tests/contrib/psycopg/test_psycopg.py
@@ -17,10 +17,8 @@ from ddtrace import Pin
 # testing
 from tests.opentracer.utils import init_tracer
 from tests.contrib.config import POSTGRES_CONFIG
-from ... import TracerTestCase
+from ... import TracerTestCase, DummyTracer
 from ...utils import assert_is_measured
-from ...utils.tracer import DummyTracer
-
 
 if PSYCOPG2_VERSION >= (2, 7):
     from psycopg2.sql import SQL
